requirements.txt
screenpy[selenium,allure,requests,appium]>=4.0.0
pytest
python-dotenv
